+++ R/boeCharts-package.R
@@ -5,11 +5,8 @@
 #' 
 #' @importFrom extrafont font_import loadfonts
 #' @import ggplot2
-#' @importFrom gtable gtable_add_grob gtable_filter gtable_add_rows gtable_trim
-#' @importFrom grid textGrob gTree gList gpar
 #' @importFrom scales dollar_format
 #' @importFrom grDevices windowsFont windowsFonts rgb col2rgb
 #' @importFrom graphics rect par image text pie
-#' @importFrom ggplotify as.ggplot
 #' @importFrom labeling extended
 "_PACKAGE"